PARTE_1/requirements.txt
autopep8==1.6.0
certifi==2021.10.8
chardet==3.0.4
click==8.0.4
Flask==2.0.3
idna==2.7
itsdangerous==2.1.1
Jinja2==3.0.3
MarkupSafe==2.1.1
marvel==0.1.0
pycodestyle==2.8.0
python-dotenv==0.19.2
requests==2.20.0
toml==0.10.2
urllib3==1.24.3
Werkzeug==2.0.3
